backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.12.5
sqlalchemy==2.0.36
python-multipart==0.0.6
pillow
python-dotenv==1.0.0
pydantic-settings==2.1.0
pyyaml
python-jose[cryptography]
passlib[bcrypt]
authlib
httpx
google-auth>=2.0.0
google-cloud-vision>=3.0.0
pytesseract>=0.3.10
